pyproject.toml
[tool.poetry]
name = "good-boi"
version = "0.1.0"
description = ""
authors = ["Andreas Rubik <rubikandreas@gmail.com"]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.0.1"
PyNaCl = "^1.5.0"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
zalgo-text = "^0.6"
scrython = "^1.11.0"
replicate = "^0.4.0"
yt-dlp = "^2022.11.11"

[tool.poetry.dev-dependencies]
ipykernel = "^6.15.1"
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
